+++ src/qq_lib/jobs/cli.py
@@ -55,6 +55,7 @@ def jobs(user: str, all: bool, yaml: bool) -> NoReturn:
             logger.info("No jobs found.")
             sys.exit(0)
 
+        BatchSystem.sortJobs(jobs)
         presenter = QQJobsPresenter(jobs)
         if yaml:
             presenter.dumpYaml()